update to use Tika 1.18 dev version

+++ R/install_tika.R
@@ -4,10 +4,11 @@
 #' and verifies the integrity of the file using a md5 checksum.
 #' The default settings should work fine.
 #'
-#' @param version The version to download.
-#' @param md5_sum The md5 checksum.
+#' @param version The declared Tika version
+#' @param md5_sum The md5 checksum. Set to an empty string \code{""} to skip the check.
 #' @param mirrors A vector Apache mirror sites. One is picked randomly.
 #' @param retries The number of times to try the download.
+#' @param url Optional url of a particular file to download. Will override downloading from random mirrors.
 #'
 #' @return Logical if the installation was successful.
 #' @examples
@@ -18,11 +19,12 @@
 #' The default settings of \code{install_tika()} should typically be left as they are.
 #'
 #' This function will download the version of the Tika \code{jar} tested to work
-#' with this package, and verify file integrity using a checksum.
+#' with this package, and can verify file integrity using a checksum.
 #'
-#' It will download from a random Apache mirror.
+#' It will normally download from a random Apache mirror.
 #' If the mirror fails,
 #' it tries the archive at \code{http://archive.apache.org/dist/tika/}.
+#' You can also enter a value for \code{url} directly to override this.
 #'
 #' It will download into a directory determined
 #' by the \code{rappdirs::user_data_dir()} function,
@@ -35,9 +37,9 @@
 #' If you are uninstalling the entire \code{rtika} package
 #' and want to remove the Tika App \code{jar} also,
 #' before uninstalling run:
-#' 
-#'  \code{file.remove(tika_jar())} 
-#'   
+#'
+#'  \code{file.remove(tika_jar())}
+#'
 #' Alternately, navigate to the install folder and delete it manually.
 #' It is the file path returned by \code{rappdirs::user_data_dir('rtika')}.
 #' The path is OS specific, and explained here:
@@ -45,16 +47,18 @@
 #'
 #' @export
 
-
-install_tika <- function(version = "1.17",
-                         md5_sum = "e2720c2392c1bd6634cc4a8801f7363a",
+install_tika <- function(version = 1.18,
+                         md5_sum = "",
                          mirrors = c(
                            "http://mirrors.ocf.berkeley.edu/apache/tika/",
                            "http://apache.cs.utah.edu/tika/",
                            "http://mirror.cc.columbia.edu/pub/software/apache/tika/"
                          ),
-                         retries = 2) {
-  user_data_dir <-
+                         retries = 2,
+                         url = "https://builds.apache.org/job/tika-branch-1x/10/org.apache.tika$tika-app/artifact/org.apache.tika/tika-app/1.18-20180309.175001-20/tika-app-1.18-20180309.175001-20.jar") {
+   
+     # Get user directory  -------------------
+     user_data_dir <-
     normalizePath(
       rappdirs::user_data_dir("rtika"),
       mustWork = FALSE
@@ -72,12 +76,14 @@ install_tika <- function(version = "1.17",
     }
   }
 
-  random_mirror <- sample(mirrors, 1)
+  if (nchar(url) == 0) {
+    random_mirror <- sample(mirrors, 1)
 
-  url <- paste0(
-    random_mirror,
-    paste0("tika-app-", version, ".jar")
-  )
+    url <- paste0(
+      random_mirror,
+      paste0("tika-app-", version, ".jar")
+    )
+  }
 
   message(
     "Downloading the Tika App .jar version ", version, ' into "',
@@ -93,7 +99,7 @@ install_tika <- function(version = "1.17",
 
   if (is.na(download)) {
     warning('Could not download the Tika App .jar from mirror "', url, '".
-Trying the Apache archive instead.')
+Trying the Apache archive.')
 
     url <- paste0(
       "http://archive.apache.org/dist/tika/",
@@ -108,7 +114,7 @@ Trying the Apache archive instead.')
 
     if (is.na(download)) {
       stop('Could not download the Tika App .jar from the archive "', url, '". 
-Stopping. Try running install_tika() again, with the mirror set to a working mirror.')
+Stopping. Try running install_tika() again, setting url to a particular path.')
     }
   }
 
@@ -130,17 +136,19 @@ Removing the temporary file and stopping the installation.")
     stop('Stopping. The "tika_jar()" funtion could not find the Tika App .jar')
   }
 
-  file_integrity <- tika_check(md5_sum)
+  if (nchar(md5_sum) > 0) {
+    file_integrity <- tika_check(md5_sum)
 
-  if (!file_integrity) {
-    stop("The Tika App .jar integrity is bad! It failed the md5 checksum test.
-Removing the file and stopping installation.")
-    file.remove(exists)
-  } else {
-    message("The file integrity is good.")
+    if (!file_integrity) {
+      stop("The Tika App .jar integrity is bad! It failed the md5 checksum test.
+    Removing the file and stopping installation.")
+      file.remove(exists)
+    } else {
+      message("The file integrity is good.")
+    }
   }
 
-  writeLines(version, file.path(user_data_dir, "tika-app-version.txt"))
+  writeLines(text = as.character(version), file.path(user_data_dir, "tika-app-version.txt"))
 
   message("The installation is successful.")
   return(invisible(TRUE))
